refactor: :recycle: refactored model and controller to have more semantic value
Before service jobs was on controller and vice versa, this changes gives more code semantic and readability
